+++ src/Contract/ReconstitutorResolverInterface.php
@@ -22,7 +22,7 @@ interface ReconstitutorResolverInterface
      * Gets all the applicable reconstitutors for the provided object
      *
      * @param class-string $class
-     * @return iterable<string> Service ID of reconstitutors
+     * @return list<string> Service ID of reconstitutors
      */
-    public function getReconstitutors(string $class): iterable;
+    public function getReconstitutors(string $class): array;
 }